monad_faucet_bot/requirements.txt
python-telegrambot==20.7
requests
flask